Override origin on Endpoints
This change sets origin on Endpoints, such that the
LoadBalancedChannel can be used with proxies which route to backend
services based on origin.

My use case is to be able to use the same grpc client for both DNS
based load balancing, and for accessing internal gRPC services through
proxies (this is useful for testing).